refactor: vec_print et mat_print affichent les float avec 2 chiffres apres la virgule et tout remis en float a la place de double

+++ src/libla/equalf.c
@@ -12,7 +12,7 @@
 
 #include "libla.h"
 
-bool	equalf(double a, double b)
+bool	equalf(float a, float b)
 {
 	return (fabs(a - b) < EPSILON);
 }
