+++ src/test/test2.c
@@ -4,11 +4,11 @@
 #define N 4
 
 // Function to perform LU decomposition
-void luDecomposition(double mat[N][N], double lower[N][N], double upper[N][N]) {
+void luDecomposition(float mat[N][N], float lower[N][N], float upper[N][N]) {
     for (int i = 0; i < N; i++) {
         // Upper Triangular
         for (int k = i; k < N; k++) {
-            double sum = 0;
+            float sum = 0;
             for (int j = 0; j < i; j++)
                 sum += (lower[i][j] * upper[j][k]);
             upper[i][k] = mat[i][k] - sum;
@@ -19,7 +19,7 @@ void luDecomposition(double mat[N][N], double lower[N][N], double upper[N][N]) {
             if (i == k)
                 lower[i][i] = 1; // Diagonal as 1
             else {
-                double sum = 0;
+                float sum = 0;
                 for (int j = 0; j < i; j++)
                     sum += (lower[k][j] * upper[j][i]);
                 lower[k][i] = (mat[k][i] - sum) / upper[i][i];
@@ -29,11 +29,11 @@ void luDecomposition(double mat[N][N], double lower[N][N], double upper[N][N]) {
 }
 
 // Function to calculate the determinant
-double determinant(double mat[N][N]) {
-    double lower[N][N], upper[N][N];
+float determinant(float mat[N][N]) {
+    float lower[N][N], upper[N][N];
     luDecomposition(mat, lower, upper);
 
-    double det = 1;
+    float det = 1;
     for (int i = 0; i < N; i++) {
         det *= upper[i][i];
     }
@@ -41,14 +41,14 @@ double determinant(double mat[N][N]) {
 }
 
 int main() {
-    double matrix[N][N] = {
+    float matrix[N][N] = {
         {3.14, 7.89, -5.67, 2.45},
         {-1.23, 4.56, 8.90, -7.34},
         {6.78, -9.01, 3.21, 5.43},
         {-4.56, 2.34, -8.76, 1.98}
     };
 
-    double det = determinant(matrix);
+    float det = determinant(matrix);
 
     printf("Determinant of the matrix is: %lf\n", det);
